use new stringshared hash, pointer was being misused since api is confusing.
The pointer type is really confusing, it was expecting pointer to
pointer what is really weird.

Cedric just added a stringshared variant that should be as fast
(removing the strcmp()) and is the api one would expect.

This fix a long standing bug I introduced (sorry!) when added the
cache, but it was bit hard to hit as if you didn't find a collision
you'd be adding to cache and never finding it. But I recently started
to use icons on desktop and for some weird reason the PDF icons
started to show in TAR :-P



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/efreet@41113 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33