proper cleanup of efreet_desktop_command.
Efreet API is BRAIN NUTS, some stuff you're supposed to release (list)
and others are up to the callback to release.

Okay, I broke e17 trying to fix efreet tests. API is so inconsistent
that its own test was wrong. I'm reverting my change and fix the test,
but this is bad.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/efreet@41133 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33